extract the MD5Stream utility from IRGen to its own header file.
